# pkg/loadbalancer: Optimize L3n4Addr.Hash for performance (2)