@aanm 2021-02-24
Changed release note since it's not a user-facing change.